Make Program into a matrix bot
Use SimpleMatrixBot to be able to tell it to book a time.
The program should now be run all the time.

It might break due to looking the thread while waiting for midnight
Add todo to make this a call back and a seperate thread.
This would make the bot part only activate this timer.